backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
httpx==0.27.2
SQLAlchemy==2.0.35
aiosqlite==0.20.0
greenlet==3.0.3
pydantic==2.9.2
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
